pyproject.toml
[tool.poetry]
name = "panotree"
version = "0.1.0"
description = ""
authors = ["Tomohiro Hayase <t.hayase@cluster.mu>", "Hikari Yanagawa <h.yanagawa@cluster.mu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
torch = [
    { version = "~=2.0.1", markers="sys_platform != 'darwin'", source = "pytorch-cuda" },  # Windows/Linux (CUDA)
    { version = "~=2.0.1", markers="sys_platform == 'darwin'", source = "PyPI"},  # macOS (CPU)
]
torchvision = [
    { version = "~=0.15.2", markers="sys_platform != 'darwin'", source = "pytorch-cuda" },  # Windows/Linux (CUDA)
    { version = "~=0.15.2", markers="sys_platform == 'darwin'", source = "PyPI" },  # macOS (CPU)
]
pyyaml = "~=6.0.1"
huggingface = "~=0.0.1"
huggingface-hub = "~=0.17.1"
timm = "~=0.9.2"
safetensors = "~=0.3.3"
numpy = "==1.25.2"
requests-toolbelt = "~=1.0.0"
pillow = "~=10.0.0"
pydantic = "~=2.6.3"
rich ="~=13.7.0"
textual = "~=0.52.1"
textual-dev = "~=1.5.1"

[[tool.poetry.source]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
